+++ modules/angular2/src/router/route_data.ts
@@ -1,4 +1,4 @@
-import {OpaqueToken} from 'angular2/src/core/di';
+import {OpaqueToken} from 'angular2/angular2';
 import {CONST_EXPR} from 'angular2/src/core/facade/lang';
 
 export const ROUTE_DATA: OpaqueToken = CONST_EXPR(new OpaqueToken('routeData'));
